Frankenstein the benchmarks
