+++ src/jmh/java/dev/blaauwendraad/masker/json/JsonMaskerBenchmark_no_streaming.java
@@ -13,8 +13,6 @@ import org.openjdk.jmh.annotations.Scope;
 import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.Warmup;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
@@ -66,10 +64,5 @@ public class JsonMaskerBenchmark {
     @Benchmark
     public byte[] jsonMaskerBytes(State state) {
         return state.jsonMasker.mask(state.jsonBytes);
-    }
-
-    @Benchmark
-    public void jsonMaskerByteArrayStreams(State state) {
-        state.jsonMasker.mask(new ByteArrayInputStream(state.jsonBytes), new ByteArrayOutputStream());
     }
 }
