+++ src/jmh/java/dev/blaauwendraad/masker/json/ValueMaskerBenchmark.java
@@ -36,12 +36,12 @@ public class ValueMaskerBenchmark {
         double maskedKeyProbability;
 
         private final JsonMasker nativeMasker = JsonMasker.getMasker(JsonMaskingConfig.builder()
-                .maskKeys("targetKey")
+                .maskKeys(Set.of("targetKey"))
                 .build()
         );
 
         private final JsonMasker rawValueMasker = JsonMasker.getMasker(JsonMaskingConfig.builder()
-                .maskKeys("targetKey")
+                .maskKeys(Set.of("targetKey"))
                 .maskStringsWith(ValueMaskers.withRawValueFunction(value -> "\"***\""))
                 .maskNumbersWith(ValueMaskers.withRawValueFunction(value -> "\"###\""))
                 .maskBooleansWith(ValueMaskers.withRawValueFunction(value -> "\"&&&\""))
@@ -49,7 +49,7 @@ public class ValueMaskerBenchmark {
         );
 
         private final JsonMasker textValueMasker = JsonMasker.getMasker(JsonMaskingConfig.builder()
-                .maskKeys("targetKey")
+                .maskKeys(Set.of("targetKey"))
                 .maskStringsWith(ValueMaskers.withTextFunction(value -> "***"))
                 .maskNumbersWith(ValueMaskers.withTextFunction(value -> "###"))
                 .maskBooleansWith(ValueMaskers.withTextFunction(value -> "&&&"))
